Refactor the bus passenger count listener, update names
